+++ CHANGES.md
@@ -1,5 +1,5 @@
 
-
+- Safe-string support. Thanks to Josh Allmann for the help.
 - Relicensed from BSD3 to ISC.
 - Build depend on topkg.
 - `uuidtrip` uses `Cmdliner` which becomes a depopt of the